moved game_loop to engine.rs
