--- src/mod.rs
@@ -1,6 +1,0 @@
-pub mod controllers;
-pub mod processors;
-pub mod structs;
-pub mod utils;
-
-
