+++ src/processors/mod.rs
@@ -1,4 +1,3 @@
 pub mod action_processor;
 pub mod conditions_processor;
-// pub mod game_generator;
 pub mod report_processor;
